정책 도메인을 패키지로 이동하라
