+++ src/main/java/com/kurly/coupon/application/factory/FactoryPolicies.java
@@ -1,6 +1,6 @@
 package com.kurly.coupon.application.factory;
 
-import com.kurly.coupon.domain.CouponPolicy;
+import com.kurly.coupon.domain.policy.CouponPolicy;
 import com.kurly.coupon.dto.CouponPolicyPublishData;
 import org.springframework.stereotype.Component;
