+++ src/main/java/com/kurly/coupon/application/factory/PolicyFactory.java
@@ -1,7 +1,7 @@
 package com.kurly.coupon.application.factory;
 
-import com.kurly.coupon.domain.CouponPolicy;
-import com.kurly.coupon.domain.PolicyType;
+import com.kurly.coupon.domain.policy.CouponPolicy;
+import com.kurly.coupon.domain.policy.PolicyType;
 import com.kurly.coupon.dto.CouponPolicyPublishData;
 
 public interface PolicyFactory {
