+++ src/main/java/com/kurly/coupon/domain/policy/CouponPolicies.java
@@ -1,4 +1,4 @@
-package com.kurly.coupon.domain;
+package com.kurly.coupon.domain.policy;
 
 import java.util.Collections;
 import java.util.List;
