+++ src/main/java/com/kurly/coupon/domain/policy/PolicyType.java
@@ -1,4 +1,4 @@
-package com.kurly.coupon.domain;
+package com.kurly.coupon.domain.policy;
 
 public enum PolicyType {
     FLAT_RATE, FIXED_AMOUNT
