+++ src/main/java/com/kurly/coupon/infra/CouponPolicyRepository.java
@@ -1,8 +1,8 @@
 package com.kurly.coupon.infra;
 
-import com.kurly.coupon.domain.CouponPolicy;
-import com.kurly.coupon.domain.Keyword;
-import com.kurly.coupon.domain.Name;
+import com.kurly.coupon.domain.policy.CouponPolicy;
+import com.kurly.coupon.domain.policy.Keyword;
+import com.kurly.coupon.domain.policy.Name;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
