+++ src/test/java/com/kurly/coupon/application/CouponServiceTest.java
@@ -1,14 +1,14 @@
 package com.kurly.coupon.application;
 
 import com.kurly.coupon.application.factory.FactoryPolicies;
-import com.kurly.coupon.domain.Amount;
-import com.kurly.coupon.domain.Count;
-import com.kurly.coupon.domain.CouponPolicy;
-import com.kurly.coupon.domain.Keyword;
-import com.kurly.coupon.domain.MinimumRedeemPrice;
-import com.kurly.coupon.domain.Name;
-import com.kurly.coupon.domain.Period;
-import com.kurly.coupon.domain.PolicyType;
+import com.kurly.coupon.domain.policy.Amount;
+import com.kurly.coupon.domain.policy.Count;
+import com.kurly.coupon.domain.policy.CouponPolicy;
+import com.kurly.coupon.domain.policy.Keyword;
+import com.kurly.coupon.domain.policy.MinimumRedeemPrice;
+import com.kurly.coupon.domain.policy.Name;
+import com.kurly.coupon.domain.policy.Period;
+import com.kurly.coupon.domain.policy.PolicyType;
 import com.kurly.coupon.dto.CouponPolicyPublishData;
 import com.kurly.coupon.infra.CouponPolicyRepository;
 import org.junit.jupiter.api.BeforeEach;
