+++ src/test/java/com/kurly/coupon/domain/policy/FixedAmountTest.java
@@ -1,4 +1,4 @@
-package com.kurly.coupon.domain;
+package com.kurly.coupon.domain.policy;
 
 import org.junit.jupiter.api.Test;
